# web/openapi.yaml
openapi: 3.0.4
info:
  title: "Social news site API"
  version: 1.0.0
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    unauthorized:
      description: "You didn't send a valid JWT"
      content:
        application/json:
          schema:
            required: [msg]
            properties:
              msg:
                example: "You must be logged in to do that!"
    posts:
      description: Success!
      content:
        application/json:
          schema:
            properties:
              posts:
                type: array
                items:
                  $ref: "#/components/schemas/post"
              count:
                type: integer
              limit:
                type: integer
              offset:
                type: integer
    adminOnly:
      description: "You must be logged in as an admin to do that!"
      content:
        application/json:
          schema:
            required: [msg]
            properties:
              msg:
                example: "You must be logged in as an admin to do that!"
    userInfo:
      description: User information
      content:
        application/json:
          schema:
            required:
              - id
              - name
              - about
              - isAdmin
              - createdAt
              - updatedAt
            properties:
              id:
                type: integer
              name:
                type: string
              about:
                type: string
              isAdmin:
                type: boolean
              createdAt:
                type: string
                format: date-time
              updatedAt:
                type: string
            format: date-time
  schemas:
    postContents:
      required:
        - title
        - category
      properties:
        title:
          type: string
        link:
          type: string
        linkType:
          type: string
        text:
          type: string
        category:
          type: string
    post:
      allOf:
        - $ref: "#/components/schemas/postContents"
        - properties:
            id:
              type: integer
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            UserId:
              type: integer
            User:
              properties:
                name:
                  type: string
            votes:
              type: integer
    postWithComments:
      allOf:
        - $ref: "#/components/schemas/post"
        - properties:
            Comments:
              type: array
              items:
                $ref: "#/components/schemas/comment"
    comment:
      properties:
        id:
          type: integer
        text:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        UserId:
          type: integer
        PostId:
          type: integer
        ParentId:
          type: integer
          nullable: true
        User:
          properties:
            name:
              type: string
        Children:
          type: array
          items:
            $ref: "#/components/schemas/comment"
          example: []
        votes:
          type: integer
  parameters:
    idInPath:
      in: path
      name: id
      schema:
        type: integer
      required: true
paths:
  /register:
    post:
      tags: [Authentication]
      summary: Create a new account
      requestBody:
        content:
          application/json:
            schema:
              required:
                - name
                - password
              properties:
                name:
                  type: string
                password:
                  type: string
                  format: password
                about:
                  type: string
      responses:
        200:
          description: "Success! You can now log in."
          content:
            application/json:
              schema:
                properties:
                  msg:
                    example: "Success! You can now log in."
        400:
          description: "Not all required fields were sent, or the username is already taken."
          content:
            application/json:
              schema:
                required: [msg]
                properties:
                  msg:
                    anyOf:
                      - example: "You must fill out the name and password fields!"
                      - example: "That name is already taken!"
  /login:
    post:
      tags: [Authentication]
      summary: Get a JWT for an account
      requestBody:
        content:
          application/json:
            schema:
              required:
                - name
                - password
              properties:
                name:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        200:
          description: "Success!"
          content:
            application/json:
              schema:
                properties:
                  msg:
                    example: "Success!"
                  token:
                    description: JWT
        400:
          description: "Not all required fields were sent."
          content:
            application/json:
              schema:
                required: [msg]
                properties:
                  msg:
                    example: "You must fill out the name and password fields!"
        401:
          description: "Wrong password!"
          content:
            application/json:
              schema:
                required: [msg]
                properties:
                  msg:
                    example: "Wrong password!"
        404:
          description: "No user exists with that name!"
          content:
            application/json:
              schema:
                required: [msg]
                properties:
                  msg:
                    example: "No user exists with that name!"
  /me:
    get:
      tags: [Users]
      summary: Get info about yourself
      security: [bearerAuth: []]
      responses:
        200: { $ref: "#/components/responses/userInfo" }
        401: { $ref: "#/components/responses/unauthorized" }
    put:
      tags: [Users]
      summary: Edit your name, password and about text
      description: The username and about text must be sent (unless you want to delete the about text). The password is only updated if a non-empty password is sent.
      security: [bearerAuth: []]
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                password:
                  type: string
                  format: password
                about:
                  type: string
      responses:
        200:
          description: Success!
        401: { $ref: "#/components/responses/unauthorized" }
  /users/{id}:
    parameters: [{ $ref: "#/components/parameters/idInPath" }]
    get:
      tags: [Users]
      summary: Get info about a specific user
      security: [bearerAuth: []]
      responses:
        200: { $ref: "#/components/responses/userInfo" }
        401: { $ref: "#/components/responses/unauthorized" }
    put:
      tags: [Users]
      summary: Edit a specific user's name, password and about text as an admin
      description: The username and about text must be sent (unless you want to delete the about text). The password is only updated if a non-empty password is sent.
      security: [bearerAuth: []]
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                password:
                  type: string
                  format: password
                about:
                  type: string
      responses:
        200:
          description: Success!
        401: { $ref: "#/components/responses/unauthorized" }
  /posts:
    get:
      tags: [Posts]
      summary: Get all posts sorted by hot
      security: [bearerAuth: []]
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        200: { $ref: "#/components/responses/posts" }
    post:
      tags: [Posts]
      summary: Create new post
      description: The link and link type; or the text field must be filled. If the link is set the link type must be set too.
      security: [bearerAuth: []]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/postContents"
      responses:
        200:
          description: Success!
        401: { $ref: "#/components/responses/unauthorized" }
  /posts/new:
    get:
      tags: [Posts]
      summary: Get all posts sorted by new
      security: [bearerAuth: []]
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        200: { $ref: "#/components/responses/posts" }
  /posts/{id}:
    parameters: [{ $ref: "#/components/parameters/idInPath" }]
    get:
      tags: [Posts]
      summary: Get a specific post and associated comments
      security: [bearerAuth: []]
      responses:
        200:
          description: Post and associated comments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/postWithComments"
        404:
          description: "No such post!"
          content:
            application/json:
              schema:
                required: [msg]
                properties:
                  msg:
                    example: "No such post!"
    delete:
      tags: [Posts]
      summary: Delete post by id
      security: [bearerAuth: []]
      responses:
        200:
          description: Success!
        403:
          description: You can only delete your own posts, unless you are an admin.
          content:
            application/json:
              schema:
                properties:
                  msg:
                    example: You can only delete your own posts, unless you are an admin.
        401: { $ref: "#/components/responses/unauthorized" }
    put:
      tags: [Posts]
      summary: Edit post by id
      security: [bearerAuth: []]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/postContents"
      responses:
        200:
          description: Success!
        403:
          description: You can only edit your own posts, unless you are an admin.
          content:
            application/json:
              schema:
                properties:
                  msg:
                    example: You can only edit your own posts, unless you are an admin.
        401: { $ref: "#/components/responses/unauthorized" }
  /topComment:
    post:
      tags: [Comments]
      summary: Add a comment to a post
      security: [bearerAuth: []]
      requestBody:
        content:
          application/json:
            schema:
              properties:
                text:
                  type: string
                PostId:
                  type: integer
      responses:
        200:
          description: Success!
        400:
          description: You can't submit a comment with no text!
        401: { $ref: "#/components/responses/unauthorized" }
        404:
          description: No such post!
  /childComment:
    post:
      tags: [Comments]
      summary: Add a child comment to a comment
      security: [bearerAuth: []]
      requestBody:
        content:
          application/json:
            schema:
              properties:
                text:
                  type: string
                ParentId:
                  type: integer
      responses:
        200:
          description: Success!
        400:
          description: You can't submit a comment with no text!
        401: { $ref: "#/components/responses/unauthorized" }
        404:
          description: No such parent comment!
  /comments/{id}:
    parameters: [{ $ref: "#/components/parameters/idInPath" }]
    delete:
      tags: [Comments]
      summary: Delete a comment
      security: [bearerAuth: []]
      responses:
        404:
          description: No such comment!
        403:
          description: You can only delete your own comments, unless you are an admin.
        200:
          description: Success!
    put:
      tags: [Comments]
      summary: Edit a comment
      security: [bearerAuth: []]
      requestBody:
        content:
          application/json:
            schema:
              properties:
                text:
                  type: string
              required: [text]
  /postVote/{id}:
    post:
      parameters: [{ $ref: "#/components/parameters/idInPath" }]
      tags: [Voting]
      summary: Add or remove vote on a post
      security: [bearerAuth: []]
      responses:
        201:
          description: Upvote added!
        200:
          description: Upvote removed!
  /commentVote/{id}:
    post:
      parameters: [{ $ref: "#/components/parameters/idInPath" }]
      tags: [Voting]
      summary: Add or remove vote on a comment
      security: [bearerAuth: []]
      responses:
        201:
          description: Upvote added!
        200:
          description: Upvote removed!
  /logs:
    get:
      tags: [Administration]
      summary: Get all log entries
      security: [bearerAuth: []]
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        200:
          description: Success!
          content:
            application/json:
              schema:
                properties:
                  logs:
                    type: array
                    items:
                      required:
                        - id
                        - path
                        - createdAt
                        - UserId
                      properties:
                        id:
                          type: integer
                        path:
                          type: string
                        createdAt:
                          type: string
                          format: date-time
                        UserId:
                          type: integer
                          nullable: true
                  count:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        401: { $ref: "#/components/responses/unauthorized" }
        403: { $ref: "#/components/responses/adminOnly" }